22/01/02 - 4-1. 패키지 재구성            	2) ui(presentation)-data(data)-domain(domain)의 패키지로 나눔
