--- app/src/main/java/com/ssafy/kurlygit/MainRepository.kt
@@ -1,6 +1,0 @@
-package com.ssafy.kurlygit
-
-interface MainRepository {
-
-    suspend fun getRepositories(search: String): Repositories
-}
