+++ app/src/main/java/com/ssafy/kurlygit/data/api/RepositoryService.kt
@@ -1,5 +1,5 @@
-package com.ssafy.kurlygit
-import retrofit2.Call
+package com.ssafy.kurlygit.data.api
+import com.ssafy.kurlygit.ui.view.data.model.Repositories
 import retrofit2.http.GET
 import retrofit2.http.Query
