+++ app/src/main/java/com/ssafy/kurlygit/data/repository/MainRepository.kt
@@ -1,0 +1,8 @@
+package com.ssafy.kurlygit.data.repository
+
+import com.ssafy.kurlygit.ui.view.data.model.Repositories
+
+interface MainRepository {
+
+    suspend fun getRepositories(search: String): Repositories
+}
