+++ app/src/main/java/com/ssafy/kurlygit/di/NetWorkModule.kt
@@ -1,5 +1,6 @@
 package com.ssafy.kurlygit
 
+import com.ssafy.kurlygit.data.api.RepositoryService
 import okhttp3.OkHttpClient
 import org.koin.dsl.module
 import retrofit2.Retrofit
