+++ app/src/main/java/com/ssafy/kurlygit/di/RepositoryModule.kt
@@ -1,5 +1,7 @@
 package com.ssafy.kurlygit
 
+import com.ssafy.kurlygit.data.repository.MainRepository
+import com.ssafy.kurlygit.data.repository.MainRepositoryImpl
 import org.koin.dsl.module
 
 val repositoryModule = module {
