+++ app/src/main/java/com/ssafy/kurlygit/di/ViewModelModule.kt
@@ -1,5 +1,6 @@
 package com.ssafy.kurlygit
 
+import com.ssafy.kurlygit.ui.viewmodel.MainViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
