+++ app/src/main/java/com/ssafy/kurlygit/ui/view/data/model/Repositories.kt
@@ -1,4 +1,4 @@
-package com.ssafy.kurlygit
+package com.ssafy.kurlygit.ui.view.data.model
 
 data class Repositories(
     val total_count :  Int,
