+++ app/src/main/res/layout/viewholder_item.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="repo"
-            type="com.ssafy.kurlygit.Repository" />
+            type="com.ssafy.kurlygit.ui.view.data.model.Repository" />
 
     </data>
